Review comments
 - Move code from nibbles.rs to stream.rs
 - Change visibility from pub(crate) to private
 - Add caution about only using it with nibbles iterators
