+++ firewood/src/nibbles.rs
@@ -140,23 +140,8 @@ impl<'a, const LEADING_ZEROES: usize> DoubleEndedIterator for NibblesIterator<'a
     }
 }
 
-pub(crate) trait IntoBytes: Iterator<Item = u8> {
-    fn nibbles_into_bytes(&mut self) -> Vec<u8> {
-        let mut data = Vec::with_capacity(self.size_hint().0 / 2);
-
-        while let (Some(hi), Some(lo)) = (self.next(), self.next()) {
-            data.push((hi << 4) + lo);
-        }
-
-        data
-    }
-}
-impl<T: Iterator<Item = u8>> IntoBytes for T {}
-
 #[cfg(test)]
 mod test {
-    use crate::nibbles::IntoBytes;
-
     use super::Nibbles;
     static TEST_BYTES: [u8; 4] = [0xdeu8, 0xad, 0xbe, 0xef];
 
@@ -264,22 +249,5 @@ mod test {
         assert_eq!(it.next(), Some(1));
         assert!(it.is_empty());
         assert_eq!(it.size_hint(), (0, Some(0)));
-    }
-
-    #[test]
-    fn remaining_bytes() {
-        let data = &[1];
-        let nib: Nibbles<'_, 0> = Nibbles::<0>(data);
-        let mut it = nib.into_iter();
-        assert_eq!(it.nibbles_into_bytes(), data.to_vec());
-    }
-
-    #[test]
-    fn remaining_bytes_off() {
-        let data = &[1];
-        let nib: Nibbles<'_, 0> = Nibbles::<0>(data);
-        let mut it = nib.into_iter();
-        it.next();
-        assert_eq!(it.nibbles_into_bytes(), vec![]);
     }
 }
